requirements.txt
yarl==1.6.3 
PyYAML==5.4.1
asyncpg==0.22.0
cryptography==3.4.7
hikari==2.0.0.dev100
tortoise-orm==0.17.2
python-dotenv==0.16.0
hikari-lightbulb==0.0.59